requirements.txt
# --- Temel Veri İşleme ---
pandas
numpy

# --- Veritabanı ---
sqlalchemy
psycopg2-binary
python-dotenv

# --- Makine Öğrenmesi Modelleri ---
scikit-learn
xgboost
lightgbm

# --- Hiperparametre Optimizasyonu ---
optuna

# --- MLOps ve Takip ---
mlflow
boto3  

# --- Scraping ---
selenium
beautifulsoup4

# --- API ve Arayüz ---
fastapi
uvicorn
streamlit
pyyaml